requirements.txt
# Core dependencies
streamlit==1.28.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3

# Additional utilities
requests==2.31.0
